requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
xlrd
lxml
google-api-python-client
google-auth-oauthlib
python-dotenv
SQLAlchemy
python-dotenv
streamlit
plotly